+++ .pre-commit-config.yaml
@@ -1,1 +1,1 @@
-/nix/store/7j0xsdz6j81avd2sycrj9kyqpy08pa23-pre-commit-config.json
+/nix/store/kgq1cqisrh2zahc3rwxgs47b2p3jdmxf-pre-commit-config.json
